Move to using Cascade class, fix the issue where the pipeline was not being updated at the same time
Commiting the whole file as I am coming back to the project after a huge break. As such, prefer to start fresh on the version control side. The main difference an new stuff is that I have now begun using the Cascade class to manage indiviual Cascade data. Additionally, a fix was made where during the creation of the Shadow Map and the update of the Cascade array, the booleans previously controlling when to update the data most likely went out of sync. As such these were removed, and the method used now is to update everything each frame. Both the shadow map and the arrays of cascades.